% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{add_primary_session}
\alias{add_primary_session}
\title{Defines the primary session}
\usage{
add_primary_session(trips = azores.fkw::trips())
}
\arguments{
\item{trips}{A tibble containinig, at least, a column named \code{date} (trip
date), whose rows are each individual trip.}
}
\value{
A tibble with a new column named \code{primary_session}.
}
\description{
\code{\link[=add_primary_session]{add_primary_session()}} adds a primary session column (\code{primary_session})
where the year is used to encode each primary session. Only trips falling
within the range of March and November get assgined a primary session
value which is the corresponding year; other dates get assigned to \code{NA}.
}
