% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{add_secondary_session}
\alias{add_secondary_session}
\title{Defines the secondary session}
\usage{
add_secondary_session(
  trips_with_prm_sessions = azores.fkw::add_primary_session()
)
}
\arguments{
\item{trips_with_prm_sessions}{A tibble with trips (each row), and containing the column \code{trip_id}.}
}
\value{
A tibble with a new column named \code{secondary_session}.
}
\description{
\code{\link[=add_secondary_session]{add_secondary_session()}} adds a secondary session column
(\code{secondary_session}) which is equal to the values in the column \code{trip_id}.
}
