% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{add_seconds}
\alias{add_seconds}
\title{Add seconds to date}
\usage{
add_seconds(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{\link[=add_seconds]{add_seconds()}} fills in the implicit zero seconds on date-times strings.
}
\examples{
add_seconds("2015-08-09 15:17")

}
