% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{add_sessions}
\alias{add_sessions}
\title{Add session annotation}
\usage{
add_sessions(
  tracks,
  psession_fn = year,
  ssession_fn = week,
  psession_var = "primary_session",
  ssession_var = "secondary_session"
)
}
\arguments{
\item{tracks}{A data frame of track data: each row should be for a track
point in space and time. The actual columns and their names may vary but the
functions passed onto \code{psession_fn} and \code{ssession_fn} need to be aware of
the available variables in \code{tracks} if that is needed to determine the
primary and secondary sessions.}

\item{psession_fn}{A function taking as first argument the \code{tracks} data
frame, whose output should be a character vector of primary session values,
whose length should match the number of rows \code{tracks}.}

\item{ssession_fn}{A function taking as first argument the \code{tracks} data
frame, whose output should be a character vector of secondary session
values, whose length should match the number of rows \code{tracks}.}

\item{psession_var}{A character value with the name of the new column storing
the primary session values, i.e. the vector resulting from a call to
the function passed in \code{psession_fn}.}

\item{ssession_var}{A character value with the name of the new column storing
the secondary session values, i.e. the vector resulting from a call to
the function passed in \code{ssession_fn}.}
}
\value{
The \code{tracks} data frame with two new columns (values are the ones
passed in \code{psession_var} and \code{ssession_var}). By default the two new
columns are \code{primary_session} and \code{secondary_session}.
}
\description{
\code{\link[=add_sessions]{add_sessions()}} adds the primary and secondary sessions to an existing data
frame of track data.

This function takes two other functions, one per session type:
\itemize{
\item \code{psession_fn}: This function should be a function taking as first argument
the data frame \code{tracks} and should return a character vector of values
indicating the primary session values. This vector length should match the
number of rows in \code{tracks}. This vector becomes a new column in the resulting
output whose name is determined by the value in \code{psession_var}.
This defaults to the \code{\link[=year]{year()}} function which looks for a date-time variable
named \code{time} and sets the primary session as the extracted year from \code{time}.
\item \code{ssession_fn}: This function should be a function taking as first argument
the data frame \code{tracks} and should return a character vector of values
indicating the secondary session values. This vector length should match the
number of rows in \code{tracks}. This vector becomes a new column in the resulting
output whose name is determined by the value in \code{ssession_var}.
This defaults to the \code{\link[=week]{week()}} function which looks for a date-time variable
named \code{time} and sets the secondary session as the extracted week from
\code{time}.
}
}
\examples{
track_example <-
  boat_trips |>
    dplyr::group_by(year) |>
    dplyr::arrange(trip) |>
    dplyr::slice(1:3) |>
    dplyr::ungroup()

track_example

# Set the primary session yearly
# and the secondary session weekly
add_sessions(track_example)

}
