% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{azores_bathymetry}
\alias{azores_bathymetry}
\title{Azores bathymetry}
\format{
\code{\link{azores_bathymetry}} is a data frame with three columns:
\describe{
\item{\code{x}}{Easting (x) coordinate, in metres, UTM Zone 26N.}
\item{\code{y}}{Northing (y) coordinate, in metres, UTM Zone 26N.}
\item{\code{depth}}{Depth, in metres.}
}
}
\usage{
azores_bathymetry
}
\description{
Bathymetry of the Azores region in long format, i.e. each row is a spatial
point.
}
\examples{
azores_bathymetry

}
\keyword{datasets}
