% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azores_islands.R
\docType{data}
\name{azores_islands}
\alias{azores_islands}
\title{Azores islands}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9 rows and 4 columns.
}
\usage{
azores_islands
}
\description{
An \link{sf} object of the nine Azorean islands. Besides the polygon geometry
it includes three fields:
\itemize{
\item \code{name}: name of the island
\item \code{group}: island group
\item \code{is_triangle}: a predicate variable that indicates whether the island is
part of the so-called of triangle islands
}
}
\examples{
azores_islands


}
\keyword{datasets}
