% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azores_triangle.R
\name{azores_triangle}
\alias{azores_triangle}
\title{Azores triangle islands}
\usage{
azores_triangle()
}
\value{
A simple feature geometry list column (\link[sf:sf]{sf}) object
containing three features, one for each of three islands.
}
\description{
Pico, Faial and São Jorge form the so-called "Triangle" of the Central Group
of islands of Azores, Portugal.
\code{\link[=azores_triangle]{azores_triangle()}} returns the islands' contours as an \link[sf:sf]{sf}
object.
}
