% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{boat_effort_for_all_trips}
\alias{boat_effort_for_all_trips}
\title{Total boat effort per grid point}
\usage{
boat_effort_for_all_trips(
  boat_effort_per_trip = azores.fkw::boat_effort_per_trip()
)
}
\arguments{
\item{boat_effort_per_trip}{A tibble as output by \code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}}.}
}
\value{
A tibble of four columns: \code{index} (index of the grid cell /
detector), UTM coordinates (\code{easting} and \code{northing}), and the total (sum)
boat effort per trip per grid point (\code{boat_effort}).
}
\description{
\code{\link[=boat_effort_for_all_trips]{boat_effort_for_all_trips()}} calculates the total boat effort per grid
point. This function takes the results per trip as generated by
\code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}} and calculates the sum of all boat efforts across
all trips per grid cell / detector.
}
\examples{
boat_effort_for_all_trips()

}
