% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{boat_effort_per_avg_trip}
\alias{boat_effort_per_avg_trip}
\title{Boat effort per average trip}
\usage{
boat_effort_per_avg_trip(
  boat_effort_per_trip = azores.fkw::boat_effort_per_trip()
)
}
\arguments{
\item{boat_effort_per_trip}{A tibble as output by \code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}}.}
}
\value{
A tibble of four columns: \code{index} (index of the grid cell /
detector), UTM coordinates (\code{easting} and \code{northing}), and the average boat
effort per trip per grid point (\code{boat_effort}).
}
\description{
\code{\link[=boat_effort_per_avg_trip]{boat_effort_per_avg_trip()}} calculates the boat effort per trip on average.
This function takes the results per trip as generated by
\code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}} and calculates the mean per grid cell / detector.
}
\examples{
boat_effort_per_avg_trip()

}
