% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_by_week.R
\name{boat_effort_per_avg_week}
\alias{boat_effort_per_avg_week}
\title{Boat effort per average week}
\usage{
boat_effort_per_avg_week(
  boat_effort_per_week = azores.fkw::boat_effort_per_week()
)
}
\arguments{
\item{boat_effort_per_week}{A tibble as output by \code{\link[=boat_effort_per_week]{boat_effort_per_week()}}.}
}
\value{
A tibble of four columns: \code{index} (index of the grid cell /
detector), UTM coordinates (\code{easting} and \code{northing}), and the average boat
effort per week per grid point (\code{boat_effort}).
}
\description{
\code{\link[=boat_effort_per_avg_week]{boat_effort_per_avg_week()}} calculates the boat effort per week on average.
This function takes the results per week as generated by
\code{\link[=boat_effort_per_week]{boat_effort_per_week()}} and calculates the mean per grid cell / detector.
}
\examples{
boat_effort_per_avg_week()

}
