% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{boat_effort_per_trip}
\alias{boat_effort_per_trip}
\title{Boat effort per trip}
\usage{
boat_effort_per_trip(
  boat_trips = azores.fkw::boat_trips_by_min,
  grid = create_grid(),
  coordinates = c("utm", "longlat")
)
}
\arguments{
\item{boat_trips}{A \link[tibble:tibble-package]{tibble} of track data (boat
trips). Data should be aggregated on a specific time interval, e.g.
minutes.}

\item{grid}{A \link[raster:raster]{RasterLayer} object.}

\item{coordinates}{The type of coordinates in the output: \code{"utm"} (default)
or \code{"longlat"}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of five plus n columns, where n
is the number of occasions. The first five columns are: \code{index}, the
detector identifier or index; \code{easting} and \code{northing}, or \code{longitude} and
\code{latitude}.
}
\description{
\code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}} generates a \link[tibble:tibble-package]{tibble}
encoding the boat effort per individual boat trip, i.e. how much time spent
on a specific grid cell or detector, per trip.
}
\examples{
boat_effort_per_trip()

}
