% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boat_trips.R
\docType{data}
\name{boat_trips}
\alias{boat_trips}
\alias{boat_trips_by_min}
\title{Boat trips}
\format{
\code{boat_trips}:
\itemize{
\item \code{year}: Year (integer).
\item \code{trip}: Name of the trip.
\item \code{time}: Date time.
\item \code{latitude}: Latitude in decimal degrees.
\item \code{longitude}: Longitude in decimal degrees.
}

\code{boat_trips_by_min}:
\itemize{
\item \code{trip}: Name of the trip.
\item \code{year}: Year (integer).
\item \code{month}: Month (integer).
\item \code{day}: Month day (integer).
\item \code{hour}: Hour in 24h format.
\item \code{min}: Minute.
\item \code{latitude}: Latitude in decimal degrees.
\item \code{longitude}: Longitude in decimal degrees.
}
}
\source{
The raw data underlying these tidy data sets is a set of GPX files
(one per year) provided as an archive in
\url{https://github.com/patterninstitute/azores.fkw/blob/main/data-raw/gpx.zip}.
}
\usage{
boat_trips

boat_trips_by_min
}
\description{
Track data of boat while performing sightseeing in the Azores' study region.
\code{\link{boat_trips}} provides a \link[tibble:tibble-package]{tibble} of the raw track
data where each row is a tracked point. \code{\link{boat_trips_by_min}} is an
aggregated version by minute. These data sets are used to define the survey
effort in this R compendium.
}
\examples{
boat_trips

boat_trips_by_min

}
\keyword{datasets}
