% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caphist_interval.R
\name{caphist_interval}
\alias{caphist_interval}
\title{Caphist interval}
\usage{
caphist_interval(x)
}
\arguments{
\item{x}{An integer vector of consecutive streaks of integer numbers. Each
streak of the same number represents a sequence of secondary sessions
belonging to the same primary session.}
}
\description{
\code{\link[=caphist_interval]{caphist_interval()}} a helper for the creation of the capthist attribute
\code{intervals}. See section \emph{Data structure and input} on
\code{vignette("openCR-vignette", package = "openCR")}.
}
\examples{
# One primary session only with three secondary sessions.
# The last primary session is always comprising of a streak of 0's in the
# output. This is the case in this interval.
caphist_interval(c(1, 1, 1))

# Two primary sessions, each with two secondary sessions.
# Note that the 1's signal the last secondary session in a primary session.
# This is true except for the last primary session, which is a streak of 0's
# without a 1 at the end.
caphist_interval(c(1, 1, 2, 2))

}
