% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{combined_effort_for_all_trips}
\alias{combined_effort_for_all_trips}
\title{Combined effort}
\usage{
combined_effort_for_all_trips(
  boat_effort = boat_effort_for_all_trips(),
  lookout_effort = lookout_effort_for_all_trips()
)
}
\arguments{
\item{boat_effort}{A tibble as returned by \code{\link[=boat_effort_for_all_trips]{boat_effort_for_all_trips()}}.}

\item{lookout_effort}{A tibble as returned by \code{\link[=lookout_effort_for_all_trips]{lookout_effort_for_all_trips()}}.}
}
\value{
A tibble with one row per spatial position, with the following
columns:
\itemize{
\item \code{index}: An index for each grid position / detector.
\item \code{easting}: Easting coordinate.
\item \code{northing}: Northing coordinate.
\item \code{longitude}: Longitude, in decimal degrees.
\item \code{latitude}: Latitude, in decimal degrees.
\item \code{lookout_score}: Lookout score, as calculated with \code{\link[=logistic_2d]{logistic_2d()}}.
\item \code{lookout_score_normalized}: Normalized lookout score, the same as
\code{lookout_score} but normalized for the total over the score over the whole
study area.
\item \code{lookout_effort}: Total lookout effort, in minutes, for the whole study
duration time.
\item \code{boat_effort}: Total boat effort, in minutes, for the whole study
duration time.
\item \code{combined_effort}: The sum of \code{lookout_effort} and \code{boat_effort}, in
minutes, for the whole study duration time.
}
}
\description{
\code{\link[=combined_effort_for_all_trips]{combined_effort_for_all_trips()}} determines the combined effort, i.e. the total of boat
and lookout efforts for all the study duration period.
}
\examples{

combined_effort_for_all_trips()

}
