% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{combined_effort_per_trip}
\alias{combined_effort_per_trip}
\title{\code{\link[=combined_effort_per_trip]{combined_effort_per_trip()}} determines the combined effort, i.e. the total of boat
and lookout efforts for every trip.}
\usage{
combined_effort_per_trip(
  boat_effort_per_trip = azores.fkw::boat_effort_per_trip(),
  lookout_effort_per_avg_trip = azores.fkw::lookout_effort_per_avg_trip()
)
}
\arguments{
\item{boat_effort_per_trip}{A tibble of boat effort per individual trip. See \code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}}.}

\item{lookout_effort_per_avg_trip}{A tibble of lookout effort per individual trip. See \link{lookout_effort_per_avg_trip}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of five plus n columns, where n
is the number of occasions. The first five columns are: \code{index}, the
detector identifier or index; \code{easting} and \code{northing}, or \code{longitude} and
\code{latitude}.
}
\description{
\code{\link[=combined_effort_per_trip]{combined_effort_per_trip()}} determines the combined effort, i.e. the total of boat
and lookout efforts for every trip.
}
