% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_by_week.R
\name{combined_effort_per_week}
\alias{combined_effort_per_week}
\title{\code{\link[=combined_effort_per_week]{combined_effort_per_week()}} determines the combined effort, i.e. the total of boat
and lookout efforts for every week.}
\usage{
combined_effort_per_week(
  boat_effort_per_week = azores.fkw::boat_effort_per_week(),
  lookout_effort_by_week = azores.fkw::lookout_effort_by_week()
)
}
\arguments{
\item{boat_effort_per_week}{A tibble of boat effort per individual week. See \code{\link[=boat_effort_per_week]{boat_effort_per_week()}}.}

\item{lookout_effort_by_week}{A tibble of lookout effort per individual week. See \link{lookout_effort_by_week}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of five plus n columns, where n
is the number of occasions (weeks). The first five columns are: \code{index}, the
detector identifier or index; \code{easting} and \code{northing}, or \code{longitude} and
\code{latitude}.
}
\description{
\code{\link[=combined_effort_per_week]{combined_effort_per_week()}} determines the combined effort, i.e. the total of boat
and lookout efforts for every week.
}
