% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_curve.R
\name{count_of_new_individuals_per_year}
\alias{count_of_new_individuals_per_year}
\title{Number of newly identified individuals per year}
\usage{
count_of_new_individuals_per_year(individuals = azores.fkw::individuals)
}
\arguments{
\item{individuals}{A tibble with the same structure as
\link[=individuals]{individuals} (this is the default).}
}
\value{
A tibble of two columns:
\itemize{
\item \code{year}: Study year, as a character.
\item \code{n}: Number of newly identified individuals (identified whale animals).
}
}
\description{
\code{\link[=count_of_new_individuals_per_year]{count_of_new_individuals_per_year()}} takes a data set of individuals
(whales) identified at a specific sighting event (\code{sighting_id}) and counts
how many individuals were newly identified each year.
}
\examples{
count_of_new_individuals_per_year()

}
\seealso{
\code{\link[=individuals_per_year]{individuals_per_year()}} \code{\link[=new_individuals_per_year]{new_individuals_per_year()}}.
}
