% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_capture.R
\name{create_capture}
\alias{create_capture}
\title{Create capture data table}
\usage{
create_capture(capture_data = create_capture_data())
}
\arguments{
\item{capture_data}{Capture data as a table, see \code{\link[=create_capture_data]{create_capture_data()}} for
format details.}
}
\value{
A table with capture data and four columns: \code{Session}, \code{ID},
\code{Occasion} and \code{Detector}.
}
\description{
\code{\link[=create_capture]{create_capture()}} takes the output from \code{\link[=create_capture_data]{create_capture_data()}} and renames
and selects columns needed for creating a capture file as required by
\code{\link[secr:DENSITY]{secr::read.capthist()}}.
}
