% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_capture.R
\name{create_capture_data}
\alias{create_capture_data}
\title{Create capture data table}
\usage{
create_capture_data(
  individuals = azores.fkw::individuals,
  trips = azores.fkw::trips(),
  grid = azores.fkw::create_grid()
)
}
\arguments{
\item{individuals}{A table listing all sighted indivuals. See \link{individuals} for format.}

\item{trips}{A table of trips. See \code{\link[=trips]{trips()}} for format.}

\item{grid}{A RasterLayer object providing the grid that is used to match
sighting locations.}
}
\value{
A table with the following columns: \code{individual}, \code{index},
\code{trip_id}, \code{primary_session}, \code{secondary_session}, \code{trip}, \code{date} and
\code{time}.
}
\description{
\code{\link[=create_capture_data]{create_capture_data()}} creates a table of unique sighting events of each
individual.
}
