% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid}
\usage{
create_grid(
  x_range = range(azores.fkw::boat_trips_by_min$longitude),
  y_range = range(azores.fkw::boat_trips_by_min$latitude),
  utm_resolution = 5000,
  x_pad_perc = 0.05,
  y_pad_perc = 0.05,
  x_offset = 0,
  y_offset = 0
)
}
\arguments{
\item{x_range}{A numeric vector of two values, specifying the minimum and
maximum values of longitude in decimal degrees.}

\item{y_range}{A numeric vector of two values, specifying the minimum and
maximum values of latitude in decimal degrees.}

\item{utm_resolution}{Grid resolution in metres.}

\item{x_pad_perc}{Padding percentage along the x-direction
(longitude/easting), i.e. how much of percentage of the \code{x_range} should
the grid be extended. This percentage is added to both ends, meaning that
effectively the \code{x_range} is increased two fold of the \code{x_pad_perc}.}

\item{y_pad_perc}{Padding percentage along the y-direction
(latitude/northing).}

\item{x_offset}{Offset along the x-direction in metres. A negative value translates
the grid location towards west.}

\item{y_offset}{Offset along the y-direction in metres. A negative value translates
the grid location towards south.}
}
\value{
A \link[raster:raster]{RasterLayer} object. By default returns a grid
encompassing the study area, i.e., off the coast of Pico, Faial and São
Jorge, Azores, Portugal.
}
\description{
\code{\link[=create_grid]{create_grid()}} creates a grid encoded as a \link[raster:raster]{RasterLayer}
object. All arguments are optional, having defaults that match the
specification required for the analysis provided by this R compendium.
}
