% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mask.R
\name{create_mask}
\alias{create_mask}
\title{Create a habitat mask}
\usage{
create_mask(
  x_range = range(azores.fkw::boat_trips_by_min$longitude),
  y_range = range(azores.fkw::boat_trips_by_min$latitude),
  utm_resolution = 1000,
  x_pad_perc = 0.05,
  y_pad_perc = 0.05,
  buffer = 12000,
  exclude_regions = NULL
)
}
\arguments{
\item{x_range}{A numeric vector of two values, specifying the minimum and
maximum values of longitude in decimal degrees.}

\item{y_range}{A numeric vector of two values, specifying the minimum and
maximum values of latitude in decimal degrees.}

\item{utm_resolution}{Grid resolution in metres.}

\item{x_pad_perc}{Padding percentage along the x-direction
(longitude/easting), i.e. how much of percentage of the \code{x_range} should
the grid be extended. This percentage is added to both ends, meaning that
effectively the \code{x_range} is increased two fold of the \code{x_pad_perc}.}

\item{y_pad_perc}{Padding percentage along the y-direction
(latitude/northing).}

\item{buffer}{Width of buffer in metres.}

\item{exclude_regions}{Regions to be excluded from the habitat mask. To be
supplied as a sf object. Coordinate units should be UTM.}
}
\value{
An object of class mask. See \link[secr:make.mask]{make.mask()} for more
details.
}
\description{
\code{\link[=create_mask]{create_mask()}} constructs a habitat mask object for spatially explicit
capture-recapture. A mask object is a set of points with optional attributes.
All arguments are optional, having defaults that match the specification
required for the analysis provided by this R compendium.
}
