% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_capture.R
\name{default_cap_header}
\alias{default_cap_header}
\title{Default capture file header}
\usage{
default_cap_header()
}
\value{
A character vector of text lines to be used as header of the capture
data file.
}
\description{
\code{\link{default_cap_header}} returns a character vector of text lines to be used
as header of the capture data file.
}
\keyword{internal}
