% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caphist_interval.R
\name{find_caphist_intervals}
\alias{find_caphist_intervals}
\title{Determine the caphist intervals}
\usage{
find_caphist_intervals(df)
}
\arguments{
\item{df}{A data frame of two columns \code{primary_session} and
\code{secondary_session}.}
}
\value{
A data frame with the same data as input \code{df} with the added
columns:
\itemize{
\item \code{primary_session_index}: An index identifying the primary session
associated with the secondary session at this row.
\item \code{caphist_interval}: A integer vector of 0's and 1's to be used for
assignment of the attribute \code{intervals} of a caphist object.
}
}
\description{
\code{\link[=find_caphist_intervals]{find_caphist_intervals()}} takes a data frame of secondary sessions (rows),
and two columns: \code{primary_session} and \code{secondary_session}; and finds the
caphist interval vector. See \code{\link[=caphist_interval]{caphist_interval()}}.
}
