% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_grid_cells.R
\name{find_grid_cells}
\alias{find_grid_cells}
\title{Find grid cells matching coordinates}
\usage{
find_grid_cells(points, grid = create_grid())
}
\arguments{
\item{points}{A \link[tibble:tibble-package]{tibble} that must contain the two
columns \code{latitude} and \code{longitude}, whose coordinates are to be matched
against the grid cells.}

\item{grid}{A \link[raster:raster]{RasterLayer} grid. Use \code{\link[=create_grid]{create_grid()}} to
create such a grid.}
}
\value{
A \link[tibble:tibble-package]{tibble} of three variables:
\describe{
\item{\code{index}}{Grid cell index.}
\item{\code{latitude}}{Latitude in decimal degrees.}
\item{\code{longitude}}{Longitude in decimal degrees.}
}
}
\description{
\code{\link[=find_grid_cells]{find_grid_cells()}} find grid cells that match the \code{points} coordinates.
}
