% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_catch.R
\docType{data}
\name{fish_catch}
\alias{fish_catch}
\title{Fish catch}
\format{
A data frame with five columns:
\describe{
\item{\code{year}}{Year.}
\item{\code{name_en}}{Fish common name in English.}
\item{\code{name_pt}}{Fish common name in Portuguese.}
\item{\code{species}}{Species name.}
\item{\code{catch}}{Fish catch in kilograms.}
}
}
\source{
\url{https://lotacor.pt/pescado-descarregado}
}
\usage{
fish_catch
}
\description{
A data set obtained from the fish market Lotaçor data portal, containing fish
catch in the Azores between 2012 and 2017 for six fish species: Thunnus
obesus, Thunnus alalunga, Katsuwonus pelamis, Thunnus albacares, Thunnus
thynnus, and Coryphaena hippurus. The \code{catch} is given in kilograms.
}
\examples{
fish_catch

}
\keyword{datasets}
