% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Import elevation data}
\usage{
get_elevation(
  file,
  resolution = 1000,
  xmin = 3e+05,
  ymin = 4100000,
  xmax = 5e+05,
  ymax = 4400000
)
}
\arguments{
\item{file}{The path to the TIFF file.}

\item{resolution}{The resolution in meters.}

\item{xmin}{Minimum easting boundary, in meters.}

\item{ymin}{Minimum northing boundary, in meters.}

\item{xmax}{Maximum easting boundary, in meters.}

\item{ymax}{Maximum northing boundary, in meters.}
}
\value{
A \link[tibble:tibble]{tibble} of five columns:
\describe{
\item{\code{x}}{Easting (x) coordinate, in metres.}
\item{\code{y}}{Northing (y) coordinate, in metres.}
\item{\code{elevation}}{Elevation. Units will depend on the units in the TIFF file
but is typically in metres.}
\item{\code{altitude}}{Same as elevation but negative values (below water surface
level) have been floored to zero.}
\item{\code{depth}}{The negative of elevation, original positive values (above
water surface level) are topped to zero.}
}
}
\description{
\code{\link[=get_elevation]{get_elevation()}} reads elevation data from a gray scale TIFF image file
for the Azores region. This function assumes that the coordinate reference
system used in the image file is \code{"+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"}. The (x, y) coordinates in the output are in UTM Zone 26N.
}
