% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_centroids.R
\name{grid_centroids}
\alias{grid_centroids}
\title{Grid centroids}
\usage{
grid_centroids(grid = create_grid())
}
\arguments{
\item{grid}{A \link[raster:raster]{RasterLayer} object. Use \code{\link[=create_grid]{create_grid()}} to
easily generate a grid.}
}
\value{
A \link[tibble:tibble-package]{tibble} of three columns: \code{index},
\code{longitude} and \code{latitude}.
}
\description{
\code{\link[=grid_centroids]{grid_centroids()}} get the centroid coordinates from a grid.
}
