% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{harmonise_date_time}
\alias{harmonise_date_time}
\title{Harmonize date and time}
\usage{
harmonise_date_time(x)
}
\arguments{
\item{x}{A character vector of date time strings.}
}
\value{
A character vector.
}
\description{
\code{\link[=harmonise_date_time]{harmonise_date_time()}} takes strings containing a date and time and returns
the same date and time always formatted in same way: \code{"YYYY-MM-DD hh:mm:ss"},
e.g. \code{"2014-05-31 17:57:35"}. It attempts to apply the following
transformations:
\itemize{
\item Uses \code{\link[=month_initials_to_indices]{month_initials_to_indices()}} to convert month initials to indices,
e.g. \code{"JUL"} to \code{"07"}. It works both with English and Portuguese month
initials, e.g. both \code{"MAI"} and \code{"MAY"} get converted to \code{"05"}.
\item Uses \code{\link[=swap_day_with_year]{swap_day_with_year()}} to convert DD-MM-YY to YYYY-MM-DD, e.g.
\code{"01-04-12"} to \code{"2012-04-01"}.
\item Uses \code{\link[=add_seconds]{add_seconds()}} to complete those strings missing the seconds components from
the time strings, e.g. \code{"15:17"} to \code{"15:17:00"}.
\item Uses \code{\link[=hour12_to_hour24]{hour12_to_hour24()}} to convert, e.g., \code{"2012-04-02 11:52:38 AM"} to
\code{"2012-04-02 11:52:38"} or \code{"01-04-12 05:54:25 PM"} to
\code{"01-04-12 17:54:25"}.
}
}
