% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{hour12_to_hour24}
\alias{hour12_to_hour24}
\title{Convert from 12 to 24-hour time}
\usage{
hour12_to_hour24(x)
}
\arguments{
\item{x}{A character vector whose elements are strings ending in the format
hh:mm:ss AM/PM.}
}
\value{
A character vector where 1:00:00 PM is converted 13:00:00, 2:00:00 PM
to 14:00:00, and so on. 12:00:00 PM remains 12:00:00. Times before noon,
i.e. 10:00:00 AM remain the same, except for the AM that is dropped.
}
\description{
\code{\link[=hour12_to_hour24]{hour12_to_hour24()}} converts time string from 12 to 24-hour time.
}
\examples{
hour12_to_hour24("2012-04-02 11:52:38 AM")
hour12_to_hour24("01-04-12 05:54:25 PM")
hour12_to_hour24("2012-10-03 12:54:08 PM")
hour12_to_hour24("2012-10-14 01:59:43 PM")
hour12_to_hour24("2013-07-06 10:33:12 AM")
}
