% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_curve.R
\name{individuals_per_year}
\alias{individuals_per_year}
\title{Individuals per year}
\usage{
individuals_per_year(individuals = azores.fkw::individuals)
}
\arguments{
\item{individuals}{A tibble with the same structure as
\link[=individuals]{individuals} (this is the default).}
}
\value{
A tibble of two columns:
\itemize{
\item \code{year}: Study year, as a character.
\item \code{n}: Number of identified individuals (identified whale animals).
}
}
\description{
\code{\link[=individuals_per_year]{individuals_per_year()}} takes a data set of individuals (whales) identified
at a specific sighting event (\code{sighting_id}) and counts how many different
individuals were identified per year.
}
\examples{
individuals_per_year()

}
\seealso{
\code{\link[=new_individuals_per_year]{new_individuals_per_year()}} for newly identified individuals per
year.
}
