% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{is_date_time}
\alias{is_date_time}
\title{Assess date time format}
\usage{
is_date_time(x)
}
\arguments{
\item{x}{A character vector of date time values to be parsed.}
}
\value{
A logical vector.
}
\description{
\code{\link[=is_date_time]{is_date_time()}} is an internal function used to assess if certain date-time variables
follow the format \code{"\%Y-\%m-\%d \%H:\%M:\%S"}.
}
\keyword{internal}
