% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_within.R
\name{is_within}
\alias{is_within}
\title{Check if points are within a polygon}
\usage{
is_within(x, y, sf_polygon)
}
\arguments{
\item{x}{A numeric vector of x-coordinates (e.g., longitude).}

\item{y}{A numeric vector of y-coordinates (e.g., latitude).}

\item{sf_polygon}{An \code{sf} object representing the polygon(s) within which the
points will be checked. The polygon object must include a column named
\code{name}.}
}
\value{
A logical vector indicating if each point is within the polygon
(\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function determines whether a set of points (defined by their
coordinates) lies within a given polygon. It returns a logical vector
indicating whether each point is within the polygon.
}
\details{
The function converts the input coordinates into an \code{sf} object and uses
\code{sf::st_within} to check whether each point lies within the provided polygon.
The \code{name} column in the \code{sf_polygon} object is used to determine membership.
Missing (\code{NA}) values in the \code{name} column indicate points that are not
within any polygon.
}
