% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secr_results.R
\name{load_secr_results}
\alias{load_secr_results}
\title{Load SECR fit results}
\usage{
load_secr_results()
}
\value{
A list of several elements:
\describe{
\item{\code{sightings}}{Sightings data frame, each observation is a sighting event.}
\item{\code{grid}}{A RasterLayer object representing the grid, the region
encompassing the study area, i.e., off the coast of Pico, Faial and São
Jorge, Azores, Portugal.}
\item{\code{effort}}{A data frame encoding the effort, i.e. how much time spent
on a specific grid cell or detector.}
\item{\code{occasions}}{A data frame of occasions from the columns in the effort data frame.}
\item{\code{capture}}{A data frame in the format expected to be saved as a capture file.}
\item{\code{habitat_mask}}{A habitat mask object for spatially explicit
capture-recapture. See \code{\link[=create_mask]{create_mask()}} for more details.}
\item{\code{sightings_loc}}{Sightings' locations, i.e. the location of the nearest grid cell centroids.}
\item{\code{fit01_params}}{SECR fit (estimated) parameters.}
\item{\code{fit01_density}}{SECR fit (estimated) density.}
\item{\code{fit01_lambda0}}{SECR fit (estimated) \eqn{\lambda_0}.}
\item{\code{fit01_sigma}}{SECR fit (estimated) \eqn{\sigma}.}
\item{\code{fit01_beta.vcv}}{TODO.}
\item{\code{fit01_details}}{TODO}
\item{\code{fit01_link}}{TODO}
\item{\code{fit01_call}}{TODO}
\item{\code{encounter_probability_fkw01_per_hour}}{TODO}
\item{\code{session_info}}{R session info.}
\item{\code{timestamp}}{Date and time of the SECR fitting.}
}
}
\description{
\code{\link[=load_secr_results]{load_secr_results()}} returns SECR fit results.
}
