% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(d, d_half = 20, rate = 0.5)
}
\arguments{
\item{d}{Distance, x-values to the logistic function.}

\item{d_half}{Distance at which the logistic function is \code{0.5}, i.e. half the
maximum y-value (that happens at \code{d = 0}).}

\item{rate}{A parameter that controls how quickly the y-values decrease over
distance.}
}
\value{
A numeric vector of y-values for the given distances (as passed
in \code{d}).
}
\description{
\code{\link[=logistic]{logistic()}} returns the logistic function y-values as a function of
distance \code{d} (x-value). The parameter \code{d_half} controls where the function
is \code{0.5}. The rate parameter tweaks how quickly the y-values decrease over
distance.
}
\examples{
logistic(d = seq(0, 40))

}
