% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{logistic_2d}
\alias{logistic_2d}
\title{Logistic function in 2D}
\usage{
logistic_2d(points, ref_point = c(-28.23645, 38.38496))
}
\arguments{
\item{points}{A numeric matrix of 2D points, where each row corresponds to a
point. First column is longitude, and second column is latitude, in decimal
degrees.}

\item{ref_point}{A numeric vector of two elements (longitude and latitude)
indicating the point of origin of the logistic function in 2 dimensions.}
}
\value{
A numeric vector of logistic y-values, one for each point provided
in the input matrix \code{points}.
}
\description{
\code{\link[=logistic_2d]{logistic_2d()}} is analogous to \code{\link[=logistic]{logistic()}} but calculates the logistic
function y-values for a set of 2D points (provided as a matrix). The origin
point can be defined with \code{ref_point}, which by default is the location of
Vigia da Queimada coordinates.
}
\examples{

p_sao_mateus <- c(-28.444493, 38.420308)
p_queimada <- c(-28.23645, 38.38496)
p_terras <- c(-28.231384, 38.385289)

points <- matrix(c(p_sao_mateus, p_queimada, p_terras), ncol = 2, byrow = TRUE)

logistic_2d(points)

}
