% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat_to_utm.R
\name{longlat_to_utm}
\alias{longlat_to_utm}
\title{Convert decimal degrees to UTM coordinates}
\usage{
longlat_to_utm(longlat, crs_from = 4326, crs_to = 32626)
}
\arguments{
\item{longlat}{A \link[tibble:tibble-package]{tibble} of points whose coordinates
are to be converted to UTM. \code{longlat} must have two columns named
\code{lon} and \code{lat}, otherwise it throws an error.}

\item{crs_from}{Source coordinate reference system (CRS), EPSG code, see the
argument \code{crs} of \link[sf:st_transform]{st_transform()} for possible
specifications; defaults to \code{4326} (WGS84).}

\item{crs_to}{Target coordinate reference system (CRS), EPSG code, see the
argument \code{crs} of \link[sf:st_transform]{st_transform()} for possible
specifications; defaults to \code{32626} (UTM zone 26N).}
}
\value{
A \link[tibble:tibble-package]{tibble} with the same number of rows as
\code{longlat}, one row per point, and two columns:
\describe{
\item{\code{northing}}{Corresponding to the conversion of latitude values.}
\item{\code{easting}}{Corresponding to the conversion of longitude values.}
}
}
\description{
\code{\link[=longlat_to_utm]{longlat_to_utm()}} converts a longitude / latitude in decimal degrees to
Universal Transverse Mercator (UTM) coordinates. The northing values are
measured from zero at the equator in a northerly direction.
}
