% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{lookout_effort_by_trip}
\alias{lookout_effort_by_trip}
\title{Lookout effort per boat trip}
\usage{
lookout_effort_by_trip(
  lookout_score_tbl = lookout_score(),
  boat_effort_per_trip_tbl = boat_effort_per_trip()
)
}
\arguments{
\item{lookout_score_tbl}{A tibble with lookout scores. See \code{\link[=lookout_score]{lookout_score()}}.}

\item{boat_effort_per_trip_tbl}{A tibble with boat effort per trip, in
minutes, see \code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}}.}
}
\value{
A tibble where boat effort in minutes has been converted to lookout
effort, also in minutes.
}
\description{
\code{\link[=lookout_effort_by_trip]{lookout_effort_by_trip()}} takes lookout scores and boat efforts per trip
in minutes, and calculates the lookout effort per trip. Essentially, the
lookout score is always the same but it is re-scaled here according to the
time the boat spent on each position. The boat trip duration is used as
a proxy for the lookout watching time.
}
