% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_by_week.R
\name{lookout_effort_by_week}
\alias{lookout_effort_by_week}
\title{Lookout effort per week}
\usage{
lookout_effort_by_week(
  lookout_score_tbl = lookout_score(),
  boat_effort_per_week_tbl = boat_effort_per_week()
)
}
\arguments{
\item{lookout_score_tbl}{A tibble with lookout scores. See \code{\link[=lookout_score]{lookout_score()}}.}

\item{boat_effort_per_week_tbl}{A tibble with boat effort per week, in
minutes, see \code{\link[=boat_effort_per_week]{boat_effort_per_week()}}.}
}
\value{
A tibble where boat effort in minutes has been converted to lookout
effort, also in minutes.
}
\description{
\code{\link[=lookout_effort_by_week]{lookout_effort_by_week()}} takes lookout scores and boat efforts per week
in minutes, and calculates the lookout effort per week. Essentially, the
lookout score is always the same but it is re-scaled here according to the
time the boat spent on each position. The boat trip duration is used as
a proxy for the lookout watching time.
}
