% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{lookout_effort_for_all_trips}
\alias{lookout_effort_for_all_trips}
\title{Lookout effort for all trips}
\usage{
lookout_effort_for_all_trips(
  lookout_score = azores.fkw::lookout_score(),
  trip_avg_time = 180,
  ntrips = ncol(boat_effort_per_trip()) - 3L
)
}
\arguments{
\item{lookout_score}{A tibble as output by \code{\link[=lookout_score]{lookout_score()}}.}

\item{trip_avg_time}{Trip average time. Default is 180 minutes.}

\item{ntrips}{Number of trips. Default is taken from the number of trips
as existing in the value returned by \code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}}.}
}
\value{
A tibble with the same structure as the value returned by
\code{\link[=lookout_score]{lookout_score()}} with an extra column: \code{lookout_effort}, i.e. the total
lookout effort corresponding to the number of trips given in \code{ntrips}.
}
\description{
\code{\link[=lookout_effort_for_all_trips]{lookout_effort_for_all_trips()}} calculates the lookout effort per trip on
average. This function takes the results generated by \code{\link[=lookout_score]{lookout_score()}} and
calculates the total lookout effort for the number of trips given in
\code{ntrips}.
}
\examples{
lookout_effort_for_all_trips()

}
