% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_by_week.R
\name{lookout_effort_for_all_weeks}
\alias{lookout_effort_for_all_weeks}
\title{Lookout effort for all weeks}
\usage{
lookout_effort_for_all_weeks(
  lookout_effort_by_week = azores.fkw::lookout_effort_by_week()
)
}
\arguments{
\item{lookout_effort_by_week}{A tibble as output by \code{\link[=lookout_effort_by_week]{lookout_effort_by_week()}}.}
}
\value{
A tibble with the same structure as the value returned by
\code{\link[=lookout_score]{lookout_score()}} with an extra column: \code{lookout_effort}, i.e. the total
lookout effort corresponding to the all weeks.
}
\description{
\code{\link[=lookout_effort_for_all_weeks]{lookout_effort_for_all_weeks()}} calculates the lookout effort per trip on
average. This function takes the results generated by \code{\link[=lookout_score]{lookout_score()}} and
calculates the total lookout effort for the number of trips given in
\code{ntrips}.
}
\examples{
lookout_effort_for_all_weeks()

}
