% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{lookout_effort_per_avg_trip}
\alias{lookout_effort_per_avg_trip}
\title{Lookout effort per average trip}
\usage{
lookout_effort_per_avg_trip(
  lookout_score = azores.fkw::lookout_score(),
  trip_avg_time = 180
)
}
\arguments{
\item{lookout_score}{A tibble as output by \code{\link[=lookout_score]{lookout_score()}}.}

\item{trip_avg_time}{Trip average time. Default is 180 minutes.}
}
\value{
A tibble with the same structure as the value returned by
\code{\link[=lookout_score]{lookout_score()}} with an extra column: \code{lookout_effort}, i.e. the lookout
effort per average trip, in minutes.
}
\description{
\code{\link[=lookout_effort_per_avg_trip]{lookout_effort_per_avg_trip()}} calculates the lookout effort per trip on average.
This function takes the results generated by
\code{\link[=lookout_score]{lookout_score()}} and calculates average lookout effort by simply multiplying
the effort score by the boat trip average duration (defaults to 180 min).
}
\examples{
lookout_effort_per_avg_trip()

}
