% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_by_week.R
\name{lookout_effort_per_avg_week}
\alias{lookout_effort_per_avg_week}
\title{Lookout effort per average week}
\usage{
lookout_effort_per_avg_week(
  lookout_effort_by_week = azores.fkw::lookout_effort_by_week()
)
}
\arguments{
\item{lookout_effort_by_week}{A tibble as output by \code{\link[=lookout_effort_by_week]{lookout_effort_by_week()}}.}
}
\value{
A tibble with the same structure as the value returned by
\code{\link[=lookout_score]{lookout_score()}} with an extra column: \code{lookout_effort}, i.e. the lookout
effort per average weekly boat trip duration, in minutes.
}
\description{
\code{\link[=lookout_effort_per_avg_week]{lookout_effort_per_avg_week()}} calculates the lookout effort per week on average.
This function takes the results generated by
\code{\link[=lookout_score]{lookout_score()}} and calculates average lookout effort by simply multiplying
the effort score by the boat weekly average duration.
}
\examples{
lookout_effort_per_avg_week()

}
