% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookout_obs_area.R
\name{lookout_obs_area}
\alias{lookout_obs_area}
\title{Lookout observation area}
\usage{
lookout_obs_area(
  lookout_max_distance = 22 * 1852,
  point_sao_mateus = c(-28.444493, 38.420308),
  point_queimada = c(-28.23645, 38.38496),
  point_terras = c(-28.231384, 38.385289)
)
}
\arguments{
\item{lookout_max_distance}{Defines the distance to the horizon, i.e. the
distance up to which the lookout is able to see. Units are in meters.}

\item{point_sao_mateus}{The position of São Mateus, Pico Island, Azores.
A numeric vector of two elements: latitude and longitude, in decimal degrees.}

\item{point_queimada}{The position of Vigia da Queimada, Pico Island, Azores.
A numeric vector of two elements: latitude and longitude, in decimal degrees.}

\item{point_terras}{The position of Terras, Pico Island, Azores.
A numeric vector of two elements: latitude and longitude, in decimal degrees.}
}
\value{
An sf object of type POLYGON.
}
\description{
\code{\link[=lookout_obs_area]{lookout_obs_area()}} returns the lookout observation area as a sf POLYGON.
This function defines the observation area as a circle section whose center
point is \code{point_queimada}, whose observation aperture range is defined by
the points \code{point_sao_mateus} to the west and \code{point_terras} to the east.
}
\examples{
(area <- lookout_obs_area())

ggplot2::ggplot() +
  ggplot2::geom_sf(data = area)

}
