% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.R
\name{lookout_score}
\alias{lookout_score}
\title{Determine lookout score}
\usage{
lookout_score(
  lookout_position = c(-28.23645, 38.38496),
  lookout_detection_fn = logistic_2d,
  lookout_obs_area_sf = lookout_obs_area(),
  grid = create_grid()
)
}
\arguments{
\item{lookout_position}{Lookout position, a numeric vector with longitude
and latitude, in decimal degrees.}

\item{lookout_detection_fn}{Lookout detection function. A function whose
x-value is the distance to the \code{lookout_position}, and whose y-value is
the "effort score".}

\item{lookout_obs_area_sf}{The observation area for the lookout, passed as
an \code{sf} object, namely of type \code{POLYGON}.}

\item{grid}{A \link[raster:raster]{RasterLayer} object. By default returns a grid
encompassing the study area, i.e., off the coast of Pico, Faial and São
Jorge, Azores, Portugal.}
}
\value{
A tibble of lookout scores for each of the grid points (as specified
in \code{grid}). The effort score is provided in two scales:
\itemize{
\item \code{lookout_score}: The score as directly calculated from the detection
function (\code{lookout_detection_fn}).
\item \code{lookout_score_normalized}: A normalized score obtained by dividing
\code{lookout_score} by the total lookout score over the study area. This variable
can be readily used and multiplied by time scaling factor, allowing the
calculation of an effective effort, e.g. if the lookout watches the sea for
180 minutes straight, i.e. for the typical time duration of a boat trip, then
the normalized score can be multiplied by 180 to determine the actual effort
per boat trip, in minutes.
}
}
\description{
\code{\link[=lookout_score]{lookout_score()}} determines the lookout "effort score" based on three
key aspects:
\itemize{
\item \code{lookout_position}: The lookout position determines the point around
which the effort score is highest. This position is by default the position
of the lookout at Vigia da Queimada, Pico Island, Azores. These coordinates
are expected in longitude and latitude, in decimal degrees.
\item \code{lookout_detection_fn}: The lookout detection function is a monotonic
decreasing function that models the effort score as a function of the
distance from the lookout position. By default this function is
\code{logistic2()}, which is the same model as used by Derville et al. (2018).
\item \code{lookout_obs_area_sf}: The lookout observation area is a polygon that
defines the region where lookout effort scores are either taken as is, or
are set to zero if outside of the region defined by this polygon. This region
is expected to be an \code{sf} object, namely of type \code{POLYGON}.
}
}
