% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{mday}
\alias{mday}
\title{Get the day of the month from date-time}
\usage{
mday(dat, dttm_var = "time")
}
\arguments{
\item{dat}{A data frame.}

\item{dttm_var}{A column name in \code{dat} identifying a date-time variable.}
}
\value{
A character vector of day of the month values.
}
\description{
\code{\link[=mday]{mday()}} uses a date-time column of \code{dat} whose name is passed in \code{dttm_var}
to extract the day of the month as character.
}
\examples{
track_example <-
  boat_trips |>
    dplyr::group_by(year) |>
    dplyr::arrange(trip) |>
    dplyr::slice(1:3) |>
    dplyr::ungroup()

track_example

# Extract day of the months from the column `time`.
mday(track_example)

}
