% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pos_by_min.R
\name{mean_pos_by_min}
\alias{mean_pos_by_min}
\title{Latitude and longitude by minute}
\usage{
mean_pos_by_min(trips = azores.fkw::boat_trips)
}
\arguments{
\item{trips}{A \link[tibble:tibble-package]{tibble} of track data, each row is a
position in time and space. By default uses \link{boat_trips}
bundled with this package.}
}
\description{
\code{\link[=mean_pos_by_min]{mean_pos_by_min()}} takes a \link[tibble:tibble-package]{tibble} with track data
(boat trips) and computes the mean latitude and longitude by aggregating
coordinates by the minute. The following columns must exist in \code{trips}:
\describe{
\item{\code{time}}{A datetime column, used for grouping by the minute.}
\item{\code{latitude}}{Latitude in decimal degrees.}
\item{\code{longitude}}{Longitude in decimal degrees.}
}
}
