% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{month_initials_to_indices}
\alias{month_initials_to_indices}
\title{Convert month initials to indices}
\usage{
month_initials_to_indices(
  x,
  initials = month.abb,
  indices = sprintf("\%02d", 1:12),
  ignore_case = TRUE
)
}
\arguments{
\item{x}{A character vector of strings containing month initials to be replaced
by indices.}

\item{initials}{A character vector of strings to be matched as month
initials. By default assumes \link{month.abb}.}

\item{indices}{A character vector of month indices. Default is \code{"01"},
\code{"02"}, etc..}

\item{ignore_case}{Whether to ignore case when matching the month initials
passed in \code{initials}. Default is \code{TRUE}.}
}
\value{
A character vector where month initials were replaced by respective
month indices.
}
\description{
\code{\link[=month_initials_to_indices]{month_initials_to_indices()}} converts month initials to indices.
}
\examples{
month_initials_to_indices(c("Jan", "Jul", "Dec"))

# By default the matching is case insensitive
month_initials_to_indices(c("JaN", "JuL", "dec"))

# If case sensitive
month_initials_to_indices(c("Jan", "JuL", "dec"), ignore_case = FALSE)
}
