% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_curve.R
\name{new_individuals_per_year}
\alias{new_individuals_per_year}
\title{Newly identified individuals per year}
\usage{
new_individuals_per_year(individuals = azores.fkw::individuals)
}
\arguments{
\item{individuals}{A tibble with the same structure as
\link[=individuals]{individuals} (this is the default).}
}
\value{
A list of newly identifier individuals. This list has as many
elements as the number of years comprising the study duration. For each
element, i.e. year, the newly identified set of individuals is provided as
a character vector of individual ids (\code{individual}).
}
\description{
\code{\link[=new_individuals_per_year]{new_individuals_per_year()}} takes a data set of individuals (whales)
identified at a specific sighting event (\code{sighting_id}) and reports the newly
identified individuals across the years.
}
\examples{
new_individuals_per_year()

}
\seealso{
\code{\link[=individuals_per_year]{individuals_per_year()}}.
}
