% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc.R
\name{perc}
\alias{perc}
\title{Percentage}
\usage{
perc(
  x,
  accuracy = NULL,
  scale = 100,
  prefix = "",
  suffix = "\%",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{accuracy}{A number to round to. Use (e.g.) 0.01 to show 2 decimal
places of precision. If NULL, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.}

\item{scale}{A scaling factor: \code{x} will be multiplied by scale before
formatting. This is useful if the underlying data is very small or very
large.}

\item{prefix}{Additional text to display before the number. The suffix is
applied to absolute value before \code{style_positive} and \code{style_negative} are
processed so that \code{prefix = "$"} will yield (e.g.) \verb{-$1} and (\verb{$1}).}

\item{suffix}{Additional text to display after the number.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{decimal.mark}{The character to be used to indicate the numeric decimal
point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common width
(see \code{base::format()}).}

\item{...}{Other arguments passed on to \code{base::format()}.}
}
\value{
A character vector of percentage formatted numbers.
}
\description{
\code{\link[=perc]{perc()}} returns numbers formatted as percentages.
}
