% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_social_network.R
\name{plot_social_network}
\alias{plot_social_network}
\title{Plot false killer whales' social network}
\usage{
plot_social_network(
  node_text_spacing = 1.1,
  sighting_colors = c(`2` = "red", `11` = "blue", `23` = "#ebcf12", `29` = "green", `30`
    = "#f700ff", `35` = "#ff6400", `43` = "#28d3d1", `44` = "#00c124", `46` = "gray", `8`
    = "black", `15` = "black", `16` = "black", `19` = "black", `20` = "black", `21` =
    "black", `25` = "black", `31` = "black", `33` = "black", `36` = "black", `37` =
    "black", `38` = "black", `40` = "black", `41` = "black", `45` = "black")
)
}
\arguments{
\item{node_text_spacing}{Use this parameter to tweak the spacing between the
node text annotation and the node.}

\item{sighting_colors}{Set the colours for each sighting.}
}
\value{
Returns a ggplot2 object that draws the aforementioned social
network.
}
\description{
\code{\link[=plot_social_network]{plot_social_network()}} uses the \code{individuals} data set to show which
whales were sighted together at each sighting event.
}
\examples{
plot_social_network()

}
