% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tracks.R
\name{read_tracks}
\alias{read_tracks}
\title{Read tracks from GPX}
\usage{
read_tracks(
  file,
  id = tools::file_path_sans_ext(basename(file)),
  id_name = "file",
  tidy_names = TRUE,
  tidy_fn = function(x) snakecase::to_snake_case(x, abbreviations = "ID")
)
}
\arguments{
\item{file}{A character vector of paths to GPX files.}

\item{id}{A character vector of values to distinguish each files' GPX track
data. By default the name of the file (without extension) is used.}

\item{id_name}{The name of the extra column to distinguish each files' GPX
track data.}

\item{tidy_names}{Whether to tidy up the column names of the output.}

\item{tidy_fn}{Function to be applied on the column names to make them tidy.
By default, it transforms the column names to snake case. This function
should take a character vector of input and return a character vector of
the same length as output.}
}
\value{
A \link[tibble:tibble-package]{tibble} of track data.
}
\description{
\code{\link[=read_tracks]{read_tracks()}} reads track data from GPX files.
}
