% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_enc_prob.R
\name{scale_enc_prob}
\alias{scale_enc_prob}
\title{Scale encounter probability}
\usage{
scale_enc_prob(p, n)
}
\arguments{
\item{p}{Base encounter probability.}

\item{n}{Period multiple factor.}
}
\description{
\code{\link[=scale_enc_prob]{scale_enc_prob()}} scales the encounter probability for a period that is \code{n}
times as big as the period that \code{p} refers to.
}
\examples{
# Assuming `p` refers to an encounter probability for a duration of 1 minute
# Then the probability over one hour (n = 60 times 1 minute) would be given
# by:
p <- 0.1
scale_enc_prob(p = p, n = 60)

# Note that increasing the period duration increases the encounter
# probability in a non-linear fashion, eventually saturating at probability
# 1 for a large `n`.

data <-
tibble::tibble(
  n = 1:100,
  p = 1-(1-0.1)^n
)

data |>
 ggplot2::ggplot(ggplot2::aes(x = n, y = p)) +
 ggplot2::geom_line() +
 ggplot2::ylim(0, 1)

}
