% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error of the mean}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Whether to remove \code{NA}.}
}
\value{
A numeric scalar.
}
\description{
\code{\link[=se]{se()}} computes the standard error of the mean.
}
