% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sightings}
\alias{sightings}
\alias{ethology}
\alias{pod_age_composition}
\alias{reaction_to_boat}
\alias{individuals}
\title{False killer whale sightings}
\format{
\code{\link{sightings}} is a data frame with eight columns:
\describe{
\item{\code{sighting_id}}{Identifier of the sighting event.}
\item{\code{latitude}}{Latitude in decimal degrees.}
\item{\code{longitude}}{Longitude in decimal degrees.}
\item{\code{date}}{Date of the sighting event.}
\item{\code{initial_time}}{Date-time of the beginning of the sighting.}
\item{\code{final_time}}{Date-time of the end of the sighting.}
\item{\code{other_species}}{Other species spotted at the same sighting event.}
\item{\code{individuals}}{Semi-colon separated list of photo-identified
individuals at each sighting.}
}

\code{\link{ethology}} is a data frame with two columns:
\describe{
\item{\code{sighting_id}}{Identifier of the sighting event.}
\item{\code{behavior}}{Whale behavior category with regards to movement,
socialization, interaction with the boat, feeding and diving.}
}

\code{\link{pod_age_composition}} is a data frame with three columns:
\describe{
\item{\code{sighting_id}}{Identifier of the sighting event.}
\item{\code{age_group}}{Age category: newborn, calf, juvenile, adult or other.}
\item{\code{n}}{Number of individuals.}
}

\code{\link{reaction_to_boat}} is a data frame with two columns:
\describe{
\item{\code{sighting_id}}{Identifier of the sighting event.}
\item{\code{reaction}}{Whale reaction towards tour boat.}
}

\code{\link{individuals}} is a data frame with eight columns:
\describe{
\item{\code{individuals}}{An identifier for each identified false killer whale.
The prefix \code{"Pc"} is the abbreviation of \emph{Pseudorca crassidens}.}
\item{\code{sighting_id}}{Identifier of the sighting event.}
\item{\code{latitude}}{Latitude in decimal degrees.}
\item{\code{longitude}}{Longitude in decimal degrees.}
\item{\code{date}}{Date of the sighting event.}
\item{\code{initial_time}}{Date-time of the beginning of the sighting.}
\item{\code{final_time}}{Date-time of the end of the sighting.}
\item{\code{other_species}}{Other species spotted at the same sighting event.}
}
}
\usage{
sightings

ethology

pod_age_composition

reaction_to_boat

individuals
}
\description{
\itemize{
\item \code{\link{sightings}} is a data set of false killer whale sighting events. At each
sighting event one or more whales may be spotted.
\item \code{\link{ethology}} is a data set of observed whale behaviors during each sighting.
\item \code{\link{pod_age_composition}} is a data set of the pod age composition for each
sighting. Each row is for a sighting/age group combination.
\item \code{\link{reaction_to_boat}} is a data set of the reactions of whales towards the
tour boat.
\item \code{\link{individuals}} is a data set of the photo-identified false killer whale
individuals. Each row is for a whale (\code{individual}) identified at a
specific sighting event (\code{sighting_id}).
}
}
\examples{
sightings

ethology

pod_age_composition

reaction_to_boat

individuals

}
\keyword{datasets}
