% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{swap_day_with_year}
\alias{swap_day_with_year}
\title{Swap day with year}
\usage{
swap_day_with_year(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{\link[=swap_day_with_year]{swap_day_with_year()}} convert a substring representing a date in the format
DD-MM-YY to YYYY-MM-DD.
}
\examples{
swap_day_with_year("01-04-12 05:54:25 PM")
swap_day_with_year("31-10-13 12:51:10 PM")

}
