% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{trips}
\alias{trips}
\title{Create a trips table}
\usage{
trips(boat_effort_per_trip = azores.fkw::boat_effort_per_trip())
}
\arguments{
\item{boat_effort_per_trip}{Boat effort per trip. See \code{\link[=boat_effort_per_trip]{boat_effort_per_trip()}}.}
}
\value{
A tibble with the following columns: \code{primary_session},
\code{secondary_session}, \code{trip_id}, \code{trip}, \code{date} and \code{time}.
}
\description{
\code{\link[=trips]{trips()}} generates a tibble where each row is for a trip.
}
