% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlat_to_utm.R
\name{utm_to_longlat}
\alias{utm_to_longlat}
\title{Convert UTM coordinates to decimal degrees}
\usage{
utm_to_longlat(utm, crs_from = 32626, crs_to = 4326)
}
\arguments{
\item{utm}{A \link[tibble:tibble-package]{tibble} of points whose coordinates
are to be converted to UTM. \code{utm} must have two columns named
\code{easting} and \code{northing}, otherwise it throws an error.}

\item{crs_from}{Source coordinate reference system (CRS), EPSG code, see the
argument \code{crs} of \link[sf:st_transform]{st_transform()} for possible
specifications; defaults to \code{32626} (UTM zone 26N).}

\item{crs_to}{Target coordinate reference system (CRS), EPSG code, see the
argument \code{crs} of \link[sf:st_transform]{st_transform()} for possible
specifications; defaults to \code{4326} (WGS84).}
}
\value{
A \link[tibble:tibble-package]{tibble} with the same number of rows as
\code{utm}, one row per point, and two columns:
\describe{
\item{\code{lon}}{Corresponding to the conversion of northing values.}
\item{\code{lat}}{Corresponding to the conversion of easting values.}
}
}
\description{
\code{\link[=utm_to_longlat]{utm_to_longlat()}} converts Universal Transverse Mercator (UTM) coordinates
to longitude / latitude in decimal degrees.
}
