% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{week}
\alias{week}
\title{Get the week from date-time}
\usage{
week(dat, dttm_var = "time")
}
\arguments{
\item{dat}{A data frame.}

\item{dttm_var}{A column name in \code{dat} identifying a date-time variable.}
}
\value{
A character vector of week values.
}
\description{
\code{\link[=week]{week()}} uses a date-time column of \code{dat} whose name is passed in \code{dttm_var}
to extract the week as character.
}
\examples{
track_example <-
  boat_trips |>
    dplyr::group_by(year) |>
    dplyr::arrange(trip) |>
    dplyr::slice(1:3) |>
    dplyr::ungroup()

track_example

# Extract weeks from the column `time`.
week(track_example)

}
