% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_fkw_probability.R
\name{whale_prob}
\alias{whale_prob}
\title{Whale encounter probability}
\usage{
whale_prob(mask, grid, capture, params, detectfn = hex)
}
\arguments{
\item{mask}{A \link[secr:mask]{mask} object containing the (x, y) locations of the
grid cells for which probabilities are to be calculated.}

\item{grid}{A \link[raster:raster]{RasterLayer} object.}

\item{capture}{A data frame of four columns: \code{Session}, \code{ID}, \code{Occasion} and
\code{Detector}.}

\item{params}{A named list containing the elements: \code{sigma} and \code{lambda0},
i.e. the fitted parameters obtained with \code{\link[secr:secr.fit]{secr.fit()}}.}

\item{detectfn}{A function object expecting three parameters: (i) \code{d}, a
double vector of distances; (ii) \code{sigma}, a scalar double of the fitted
\eqn{\sigma}; and (iii) \code{lambda0}, a scalar fitted \eqn{\lambda_0}.}
}
\value{
A \link[tibble:tibble]{tibble} that is the \code{mask} object passed as
input with the added column \code{p} for the encounter probability.
}
\description{
\code{\link[=whale_prob]{whale_prob()}} determines whale encounter probabilities.
}
