% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_min_duration.R
\name{which_min_duration}
\alias{which_min_duration}
\title{Where is the minimum duration}
\usage{
which_min_duration(duration)
}
\arguments{
\item{duration}{A numeric vector.}
}
\value{
A logical vector.
}
\description{
\code{\link[=which_min_duration]{which_min_duration()}} takes a numeric vector of durations and returns
the minimum (strictly) positive duration.
}
