% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_week_of_study.R
\name{which_week_of_study}
\alias{which_week_of_study}
\title{Find study week for query dates}
\usage{
which_week_of_study(dttm)
}
\arguments{
\item{dttm}{A date or date-time vector of points in time to be assigned
a week of study.}
}
\value{
A character vector of elements of the form \code{"week n"}, where \code{n} is
the number of the week. Note that \code{"week 1"} is the first week of 2012, and
week \code{"week 313"} is the last week of 2017.
}
\description{
\code{\link[=which_week_of_study]{which_week_of_study()}} finds the week of study to which the queried dates
(\code{dttm}) match. So this function allows us to assign a week to specific
dates.

By week of study we mean the number of the week starting in 2012 until the
end of 2017.
}
