% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_capture.R
\name{write_capture}
\alias{write_capture}
\title{Write capture to file}
\usage{
write_capture(x, file, header = default_cap_header())
}
\arguments{
\item{x}{A \link[tibble:tibble-package]{tibble} containing capture (secr) related
information, namely the four columns: \code{Session} (block of sampling that may
be treated as independent), \code{ID} (individual identifier), \code{Occasion}
(occasion identifier) and \code{Detector} (detector identifier). Each row
identifies correspond to an individual (\code{ID}) seen at a certain time
(\code{Occasion}) on a certain location (\code{Detector}).}

\item{file}{The path to a file to saved.}

\item{header}{A character vector to be used as header in the file. Do not
include the \code{"#"} as these are added automatically to every line. Defaults to
the value returned by \code{\link[=default_cap_header]{default_cap_header()}}.}
}
\value{
Returns the path passed in \code{file}.
}
\description{
\code{\link[=write_capture]{write_capture()}} writes a capture \link[tibble:tibble-package]{tibble} to a text
file.
}
