% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_effort.R
\name{write_effort}
\alias{write_effort}
\title{Write effort to a file}
\usage{
write_effort(
  x,
  file,
  detector_col = "index",
  x_col = "easting",
  y_col = "northing"
)
}
\arguments{
\item{x}{A \link[tibble:tibble-package]{tibble} encoding the effort. See the
function \code{\link[=effort]{effort()}} on how to generate this tibble.}

\item{file}{A path to a file to be exported.}

\item{detector_col}{Name of the detector column.}

\item{x_col}{Name of the column indicating the X position.}

\item{y_col}{Name of the column indicating the Y position.}
}
\value{
Returns the path passed in \code{file}.
}
\description{
\code{\link[=write_effort]{write_effort()}} writes a text file with a \link[tibble:tibble-package]{tibble}
representing the effort.
}
