% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{yday}
\alias{yday}
\title{Get the day of the year from date-time}
\usage{
yday(dat, dttm_var = "time")
}
\arguments{
\item{dat}{A data frame.}

\item{dttm_var}{A column name in \code{dat} identifying a date-time variable.}
}
\value{
A character vector of day of the year values.
}
\description{
\code{\link[=yday]{yday()}} uses a date-time column of \code{dat} whose name is passed in \code{dttm_var}
to extract the day of the year as character.
}
\examples{
track_example <-
  boat_trips |>
    dplyr::group_by(year) |>
    dplyr::arrange(trip) |>
    dplyr::slice(1:3) |>
    dplyr::ungroup()

track_example

# Extract day of the years from the column `time`.
yday(track_example)

}
