% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{year}
\alias{year}
\title{Get the year from date-time}
\usage{
year(dat, dttm_var = "time")
}
\arguments{
\item{dat}{A data frame.}

\item{dttm_var}{A column name in \code{dat} identifying a date-time variable.}
}
\value{
A character vector of year values.
}
\description{
\code{\link[=year]{year()}} uses a date-time column of \code{dat} whose name is passed in \code{dttm_var}
to extract the year as character.
}
\examples{
track_example <-
  boat_trips |>
    dplyr::group_by(year) |>
    dplyr::arrange(trip) |>
    dplyr::slice(1:3) |>
    dplyr::ungroup()

track_example

# Extract years from the column `time`.
year(track_example)

}
